requirements.txt
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pymssql==2.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
tqdm==4.51.0
